+++ CHANGES.md
@@ -5,6 +5,7 @@
 - Corrected typographical errors in the README.md file.
 - Corrected the repository in the package.json file.
 - Corrected comments in the GraphAPI.js file.
+- The callback data is the response.value and not the response itself.
 
 ## Version 0.0.1